requirements.txt
crewai[google-genai]
streamlit
cognee
python-dotenv
crewai-tools